23-05-22 작업(1)

+++ data/src/main/kotlin/com/dmonster/data/remote/dto/MemberInfoDto.kt
@@ -4,7 +4,7 @@ import com.dmonster.data.remote.dto.base.BaseResponse
 import com.google.gson.annotations.SerializedName
 
 data class MemberInfoDto(
-    @SerializedName("idx") val idx: Int,
+    @SerializedName("idx") val idx: Long,
     @SerializedName("mt_id") val mt_id: String? = null,
     @SerializedName("mt_name") val mt_name: String? = null,
     @SerializedName("mt_wdate") val mt_wdate: String? = null,
