+++ domain/src/main/kotlin/com/dmonster/domain/model/MemberInfoModel.kt
@@ -3,12 +3,12 @@ package com.dmonster.domain.model
 import com.dmonster.domain.model.base.BaseModel
 
 data class MemberInfoModel(
-    val idx: Int,
+    val idx: Long,
     val mt_id: String? = null,
     val mt_name: String? = null,
     val mt_wdate: String? = null,
     val mt_type: Int? = null,
 ) : BaseModel {
-    override val key: Int
+    override val key: Long
         get() = idx
 }
