+++ domain/src/main/kotlin/com/dmonster/domain/model/navi/BottomNavItem.kt
@@ -3,10 +3,10 @@ package com.dmonster.domain.model.navi
 import com.dmonster.domain.model.base.BaseModel
 
 data class BottomNavItem(
-    val icon: Int,
+    val icon: Long,
     val title: String,
     var isChecked: Boolean,
 ): BaseModel {
-    override val key: Int
+    override val key: Long
         get() = icon
 }
